New folder/requirements.txt.txt
streamlit
pandas
rembg
pillow
openpyxl